Update: Added layouts for the main pages. Also, added some pictures/videos that might be used in the simulation page
The whole purpose of creating multiple pages is to allow multiple people to revise/edit and add onto different sections. There are many sections, so this should allow for a lot of work to be done simultaneously. I will be working on a cleaner looking page design, but this will do for now. I will also be editing these photos, but wanted to add them onto here so I could easily find them again. The own-cloud page needs to be cleaned up as so many people are posting photos and stuff in so many different sections I get lost every time I look for some. I'll have to fix some of that over the weekend.